@@ -13,7 +13,7 @@ omit = [
 ]
 
 [tool.black]
-target-version = ["py38"]
+target-version = ["py39"]
 
 [[tool.mypy.overrides]]
 module = [
@@ -47,7 +47,7 @@ log_level = "DEBUG"
 
 [tool.ruff]
 line-length = 88
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 ignore = []
