@@ -23,7 +23,7 @@ packages =
     fuzzing_decision.pool_launch
 include_package_data = True
 zip_safe = False
-python_requires = >=3.8
+python_requires = >=3.9
 
 [options.extras_require]
 decision =
