@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.black]
-target-version = ["py38"]
+target-version = ["py39"]
 
 [[tool.mypy.overrides]]
 module = [
@@ -17,7 +17,7 @@ max-line-length = 88
 
 [tool.ruff]
 line-length = 88
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 select = [
