@@ -13,7 +13,7 @@ install_requires =
     fasteners
 py_modules =
     rwait
-python_requires = >=3.8
+python_requires = >=3.9
 
 [options.entry_points]
 console_scripts =
