@@ -15,7 +15,7 @@ install_requires =
 py_modules =
     nyx_ipc_manager
     nyx_utils
-python_requires = >=3.8
+python_requires = >=3.9
 
 [options.entry_points]
 console_scripts =
