@@ -15,7 +15,7 @@ package_dir =
 packages =
     orion_builder
 zip_safe = False
-python_requires = >=3.8
+python_requires = >=3.9
 
 [options.entry_points]
 console_scripts =
