@@ -11,7 +11,7 @@ from os import getenv
 from pathlib import Path
 from platform import machine
 from shutil import rmtree
-from typing import List, Optional
+from typing import Optional
 
 from taskboot.build import build_image
 from taskboot.target import Target
@@ -110,7 +110,7 @@ class BuildArgs(CommonArgs):
             )
 
 
-def main(argv: Optional[List[str]] = None) -> None:
+def main(argv: Optional[list[str]] = None) -> None:
     """Build entrypoint. Does not return."""
     args = BuildArgs.parse_args(argv)
     base_tag = "latest"
