@@ -6,7 +6,7 @@
 from pathlib import Path
 from subprocess import CalledProcessError
 from tempfile import gettempdir
-from typing import Dict, Optional
+from typing import Optional
 from unittest.mock import call
 
 import pytest
@@ -214,8 +214,8 @@ def test_refs() -> None:
 def test_github_tc(
     mocker: MockerFixture,
     action: str,
-    event: Dict[str, Dict[str, str]],
-    result: Dict[str, Optional[str]],
+    event: dict[str, dict[str, str]],
+    result: dict[str, Optional[str]],
     repo_args: str,
 ) -> None:
     """test github event parsing from taskcluster"""
